feat: Lookup outputToken symbol using TOKEN_SYMBOLS_MAP
Currently we map a deposit's outputToken to an L1TokenInfo object via the HubPoolClient. This will no longer work once outputTokens are supported that are not mapped to the HubPool via PoolRebalanceRoutes.

This PR replaces the hubPoolClient. getL1TokenInfoForL2Token call wherever the outputToken is supplied with a `getL1TokenInfo` call that looks into TOKEN_SYMBOLS_MAP

We replace this in a few spots:
- Relayer.handleUnprofitableFill which logs about unprofitable deposits.
- Relayer.handleTokenShortfall which maps shortfalls with outputTokens
- ProfitClient which looks up token prices of output tokens

In addition, this PR adds a hard-coded symbol remapping in the ProfitClient that forces price equivalency between USDC.e, USDbC and USDC symbols, which will be required to support USDC -- USDC.e/USDbC deposits post-CCTP migration.