feat: Lookup outputToken symbol using TOKEN_SYMBOLS_MAP (#1461)
* feat: Lookup outputToken symbol using TOKEN_SYMBOLS_MAP

Currently we map a deposit's outputToken to an L1TokenInfo object via the HubPoolClient. This will no longer work once outputTokens are supported that are not mapped to the HubPool via PoolRebalanceRoutes.

This PR replaces the hubPoolClient. getL1TokenInfoForL2Token call wherever the outputToken is supplied with a `getL1TokenInfo` call that looks into TOKEN_SYMBOLS_MAP

We replace this in a few spots:
- Relayer.handleUnprofitableFill which logs about unprofitable deposits.
- Relayer.handleTokenShortfall which maps shortfalls with outputTokens
- ProfitClient which looks up token prices of output tokens

In addition, this PR adds a hard-coded symbol remapping in the ProfitClient that forces price equivalency between USDC.e, USDbC and USDC symbols, which will be required to support USDC -- USDC.e/USDbC deposits post-CCTP migration.

* move getL1Token util fnuction into HubPoolClient

* Update ProfitClient.ts

* Fix

* Fix

* Remove reliance in getL1TokenInfo on there being an Mainnet entry for l2 token

* Update Relayer.ts

* Update Relayer.ts

* Call hub pool client directly to get l2 token mapped to l1 token when computing refund/running balance data

* Import TOKEN_EQUIVALENCE_REMAPPING from constants-v3

* Update InventoryClient.ts

* Update InventoryClient.ts

* Update InventoryClient.ts

* Update InventoryClient.ts

---------

Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>