improve(relayer): Extend deposit confirmation validation
This change makes the relayer more responsive to unfilled deposits
without compromising on safety.

It achieves this by filtering each deposit individually for the minumum
number of required confirmation requirements before computing the batch
deposit confirmation requirements for set of outstanding deposits. This
ensures that individual large deposits that have not met the minimum
required threshold cannot DoS lower-value deposits that have met their
minimums. In hindsight, this logic is completely obvious.

In order to avoid repeatedly sorting and filtering on the deposit
confirmation structure, change its format to be a sorted array. This
makes it less computationally-expensive (and less annoying) to evaluate
MDC requirements.