improve(TokenClient): multicall balance & allowance queries (#1496)
TokenClient update is occasionally observed to hang for 10s of seconds.
It's a surprisingly RPC-intensive method because it issues a parallel
RPC query for each supported token on a chain. This doesn't scale well
with each new chain or token that's added, so bundle the queries into a
multicall. The legacy implementation is retained for chains that don't
support multicall3, which is nominally just hardhat in test.

To simplify this change, split out the HubPool BondToken query from the 
regular update() call. The BondToken part is so specific that it could
optionally be pulled out of the TokenClient entirely.