improve(reporter): Add special USDC.e entry
Adds some special logic to break out USDC.e balances from USDC. L1 USDC is counted with Native USDC balance.

This is required to be a special case because the balance report uses the L1 token symbol mapped to all L2 tokens as the key, whereas USDC.e and Native USDC on L2s both map to a single USDC token, so we need to invent a "USDC.e" virtual L1 token to keep the structure of the report intact.

A more scalable change would have to change the report structure fundamentally.