improve(relayer): Persist fillStatus across loops (#1588)
The looping relayer is forced to drink from the fire hose when
determining the set of unfilled deposits. In times of high demand this
can lead to a lot of unnecessary looping to filter out deposits that
have already been filled. This change adds tracking of fill status
across loops in order to permit already-filled deposits to be filtered
out as early as possible.