fix(Dataworker): Deep clone refunds object when constructing RelayerRefundRoot
In `DataworkerUtils`, when we construct the `combinedRefunds` object using the `refunds` from `bundleFillsV3` (i.e. the fast fill refunds in a bundle) and the `expiredDepositRefunds` (i.e. refunds for expired deposits), we accidentally modify the former as a reference variable.

This leads to upstream issues in `proposeRootBundle` which returns the bundle data to post to Arweave [here](https://github.com/across-protocol/relayer/blob/13c5f17db14f11789f3fb6322f52183293fdb3cd/src/dataworker/Dataworker.ts#L488). However, the very subtle bug is that the values we save to post to arweave here are the same references we accidentally mutate above.

This PR fixes the bug and also ensures we post cloned objects to arweave to reduce chance of mutations