fix(relayer): Skip cross-chain transfer tracking when looping (#1621)
Checking the logs more closely, it's obvious that some chain adapters
repeatedly produce the same copies of deposit hashes. This creates
duplicates and will result in delays and odd results in balance
projections. A proper fix is needed in the adapters, but this is
deferred for now because of the ongoing new-chain automation work.