fix(AcrossAPIClient): Support liquid-reserves query (#1636)
The API was recently updated to return a modified definition of
maxDeposit, such that it is capped at the largest desirable deposit
amount (but not necessarily the largest deposit that Across can 
support). This change also added a destinationChainId-dependent aspect,
which is incompatible with the relayer's existing use of the API where
it tries to only query from anywhere -> mainnet to determine limits.
This caused flow-on effects in the relayer where it would refuse to make
fills that were perfectly OK to fill. This change restores the ability
of the relayer to accurately fill based on configured limits.

---------

Co-authored-by: nicholaspai <9457025+nicholaspai@users.noreply.github.com>