fix(finalizer): define contract addresses required for Lisk finalizations (#1662)
The optimism SDK does not have predefined contract addresses for Lisk
like it does for all of our other OP Stack chains. This causes
finalizations to throw undefined errors. This commit adds in the
necessary contract addresses which we then pass into the optimismSDK's
CrossChainMessenger. We will also need to do something similar for
Blast.

---------

Signed-off-by: bennett <bennett@umaproject.org>