improve(relayer): Avoid friendly fire when new chains are added
The relayer currently has an annoying constraint that it needs to be
updated each time a new chain is added to the config store. This is
because the post-update configuration validation is unnecesarily strict.
In cases where the relayer has a custom set of origin and destination
chains, it actually doesn't make sense to require that configuration is
defined for any new chain that has been added.

So:
- When the relayer does not rely on a new chain, don't verify
  configuration for that chain.
- When the relayer does rely on a new chain, warn about missing config
  and revert to a conservative default of 5k blocks for eth_getLogs()
  calls. The warnings should prompt the operator to update their bot or
  define the necessary config in their environment.