fix(Blast): Correct WETH address (#1672)
Blast uses a non-standard OP address for WETH. The address we had 
specified was off by a single digit (0x42.....04 instead of 0x43..04).
0x42..04 is an OP predeploy, so deposit() simulation failed. Had it been
an EOA then there was potential for loss of ETH when attempting to wrap
to WETH. To protect against this in future, first call WETH.symbol() to
verify that the destination contract is an ERC20, and secondly to verify
that the contract claims to be a WETH implementation. This should catch
the absolute majority of mistakes that could occur here.