refactor(test): Split Across v3 BundleDataClient tests (#1663)
This single test file takes ~10 minutes to execute in CI and is by far
the longest running test. CI can be sped up by splitting part of this
test out into a separate file, such that they can be executed in
parallel. This reduces the runtime for this set of tests by
approximately 50%, and shaves ~3 minutes of the entire test run.

While here, apply a couple of minor type corrections.