refactor(relayer): Simplify deposit confirmation config (#1546)
DEFAULT_MIN_DEPOSIT_CONFIRMATIONS isn't actually used anymore, so drop
it, but migrate the chain finality defaults into
MIN_DEPOSIT_CONFIRMATIONS. Also, split out testnet chain IDs to be
dynamically generated based on predefined chains, but only when the 
relayer is not running on mainnet. This should make it a bit simpler to
update and maintain in future.