improve(InventoryClient): 1x overage on lite chains (#1683)
Lite chains will have a tendency to creep over target and require manual
rebalancing. Additionally, taking voluntary repayment on a lite chain is
actually a net cost to the dataworker because it requires tokens to be
bridged to the lite chain, so it causes higher gas consumption for the 
deposit transactions on mainnet. Try to limit this by clamping the 
default overage buffer on lite chains to 1x, instead of the existing
default 1.5x.