fix(ProfitClient): Fix fill amount USD computation
The ProfitClient currently assumes that the value of a fill can be
determined from the input token price * the output amount of the
output token. This is roughly correct for most Across deposits, but
falls over when the deposit includes an in-protocol swap to an output
token with a different number of decimals.

This was exposed on an Optimism deposit for inputToken USDC, outputToken
AERO on Base.

Deposit hash:
  0x4a3febcfb764467d9eae261b3a3938137db383289fe328050a2871eb407e1147