fix(executor): Force no multicall on Polygon
Relayer refund leaves must not be batched via multicall on Polygon. The
existing executor implementation will attempt multicall in cases where
there are multiple relayer refund leaves, and this imposes a soft
deadlock on the executor. This can be recovered by executing one leaf at
a time. A better solution is just to override the defaults for Polygon
in the dataworker.