fix(InventoryClient): Subtle target overage syntax error
The existing one-liner was marginally too clever and resulted in
unintuitive behaviour. In the pre-existing code the expression would
unconditionally be evaluated via the ternary statement, so any
env-defined target overage would be ignored.

It's apparently not valid (...or maybe not wise) to combine a nullish
coalescing operator with a ternary statement. It is possible to fix this
in a one-liner by wrapping the ternary statement in brackets, but it
feels risky to retain that so I've split the expression into two lines.