Require manual env config
Handling the defaults is tricky because targetOverageBuffer is
unconditionally set in the relayer config, at the point where no
information about lite chains is available. Relayer config also has no
interface to query this information.

To keep things simple, just require that the operator configure overages
manually in the event that they're needed for a lite chain. Maybe this
can be revisited later.