improve(relayer): More robust fast relayer event ingestion
The fast relayer currently imposes a finality overlay onto each indexer
process, such that each process will wait for finality blocks, and will
then forward events that have met quorum, or discard them. This is
fragile on chains with very short block times and results in some events
being dropped due to delayed delivery of events by some RPC providers.
Additionally, on chains with a high number of MDCs, it imposes an
arbitrary delay on relaying the events. This can be meaningful in the
case of FilledV3Relay events, which are now used by the relayer to track
its origin chain commitments.

With this change, retain events for a longer period and forward them if
and when they reach the minimum configured quorum. This is much more
robust and should result in a greater number of events making it through
to the relayer.

A specific concern with this change was the potential for events to be
ingested "out of order". The impact of this seems to be very low or
non-existent, with the observation that the SpokePoolClient tends to
store events in mappings and sorts them on demand.