fix: Try to lessen Linea finalizer RPC burden
The Linea SDK implements some unfortunate RPC queries that are not
well-suited to being called in parallel. This seems to be resulting in
errors being returned by upstream RPC providers. Serialise some queries,
with a longer delay on retry, in order to try to lessen the load and
reduce the error rate. This will inevitably lead to a longer execution
time, but might help to save the bot from throwing.