improve(relayer): Support dynamic handover in production
The fast relayer can presently have multiple concurrent instances
for a short period of time because there is no method of coordinating
the "active" relayer. This can cause some collisions when both relayer
instances attempt to fill the same transaction(s), and requires that the
relayer be stopped via a SIGHUP, which is slightly awkward in the
serverless environment.

This change instead uses redis as a coordination mechanism between
relayer instances, such that a new relayer will update a specific redis
record when it is ready to take over from any previous instance.
Likewise, and existing instance will detect when the record is updated
and will gracefully exit thereafter.