improve(relayer): Test lowered ethers pollingInterval
The current ethers polling interval is 4 seconds. In case of waiting on
transaction confirmation, this probably imposes unnecessary delays in
the fast relayer when it has many fills to make.

This is done for test purposes to see whether relayer fill times improve
as a result. Subject to the results, this should be upstreamed and made
configurable, such that the fast relayer can opt-in to a lower
pollingInterval, and that it can tune the interval per chain.