improve(relayer): Support keepalives over websockets (#1771)
As newer chains are added, we sometimes see that websockets are
occasionally dropped; potentially as a result of low activity on the
chain meaning that the websocket sits idle. This commit adds a simple
periodic keepalive that will be initiated by each listener process
towards each of its configured providers.