refactor(relayer): Add one-time relayer init
This ensures that token approvals are in place for all tokens held by
the relayer EOA. For 99.999% of the time it's fine to do this once on
startup; it doesn't need to be repeated on each loop.

This also removes a redundant tokenClient.update() call that is being
performed on each loop.