refactor(relayer): Rate-limit maintenance in looping mode
There's no need to compulsively check ETH and WETH balances in looping
mode. This clutters the logs a lot, increases the relayer's exposure
to delays on RPC queries and generally unnecessarily increases RPC
consumption. Instead, rate-limit these taks to a default of 1/minute,
with the option to override via the RELAYER_MAINTENANCE_INTERVAL
configurable .