fix(relayer): Don't prematurely mark deposits as filled
Per current implementation, the relayer can prematurely mark a deposit
as filled if it _intends_ to fill it. This can however be derailed in
the event that the destination chain has pending fills, in which case
the fill transaction is discarded. This creates problems for fill status
tracking, which is otherwise marked at the time the transaction is
enqueued.

The simple fix is to filter candidate deposits by whether their
destination chain has pending fills. A more optimised version, which
might follow, would apply a pending status for these fills and would
apply that pending status at the time of transaction submission. That's
more complex for an (at best) marginal benefit, so proceed with the
simple solution for now.