across: Update token balances unconditionally on token shortfalls
The relayer currently only updates its token balances every x seconds
(nominally 60). This is a nice optimisation for the standard case, but
results in unnecessary marginal delays in the case of proactive
rebalancing or awaiting on repayments. Instead, modify the update logic
to be conditional on token shortfalls OR at least x seconds (nominally 60).