improve(arweave): Change arweave data key from list of chain ID's to map
We've received feedback from a user of the arweave data that its easier if the arweave `tag` (i.e. data key) is a mapping of chainId to bundle block ranges. I see no issue with this outside of any topic length constraints, which I've added logic to check for and protect against. The tag now only uses the end blocks, to reduce its length.