improve(Relayer): Add total gas price to `info` log when relayer sends fill
This will help debugging and give us broad oversight into how accurate our gas price estimates are. We probably should refactor the overall relay fee calculation flow but this PR uses the data we already have (native and token gas costs) to derive the gas price easily.

This PR also adds to the ProfitClient's `updateGasCosts` function a simple change to print out the gas price estimate per chain.

A more helpful fix might be to add the `gasPrice` field to the `TransactionCostEstimate` type in the SDK and force functions like `sdk.common.estimateTotalGasRequiredByUnsignedTransaction()` to return the priority fee and base fee broken down so the relayer can log it