improve(Constants): Update default deposit confirmations (#1968)
The existing defaults are set so conservatively that third-party
operators are pushed into overriding them with custom values. This
introduces a greater risk of loss of funds. Instead, update the
defaults with lower confirmation requirements that should permit new
operators to make fills successfully without having to discard the
defaults.