fix(Dataworker): Cannot fund DonationBox using MulticallerClient (#1975)
* fix(Dataworker): Cannot fund DonationBox using MulticallerClient

The MulticallerClient has [logic](https://github.com/across-protocol/relayer/blob/40b0c7971a3548a0499e7fc5990d150437c527fb/src/clients/MultiCallerClient.ts#L361) that batches transactions together and attempts to `multicall()` them. This works normally with the `HubPool` and `SpokePool` which extend the `Multicaller()` interface.

However, we can't use this logic to make calls on ERC20's like we are doing to fund the `DonationBox`. Instead, we either need to update the `MulticallerClient` to handle cases where transactions are not multicall-able, or we can send these transactions to the `TransactionClient`.

In this PR, I added a new method `simulateAndSubmit` to the `TransactionClient` that the Dataworker uses to `transfer` custom gas tokens to the DonationBox.

* Update src/dataworker/Dataworker.ts

* Update Dataworker.executePoolRebalanceUtils.ts

* Update TransactionClient.ts

* Update TransactionClient.ts

* Update src/clients/TransactionClient.ts

Co-authored-by: James Morris, MS <96435344+james-a-morris@users.noreply.github.com>

* Use Multicaller client

* Consolidate valueTxnsand nonMulticallTxns

* Update MultiCallerClient.ts

* Update MultiCallerClient.ts

---------

Co-authored-by: James Morris, MS <96435344+james-a-morris@users.noreply.github.com>