improve(relayer): Reduce token swap log spam (#2007)
This log message was added during the v2 -> v3 transition where it was
interesting to know about deposits with unsupported outputTokens. There
are now other protocols building on top of Across and supporting this,
so it generates quite a bit of log spam. Reduce the loglevel and the
information bundled in the log.