improve(l1l2Finalizer): Make sure L1 finalizer uses hub chain spokepool's lookback
Currently we are always using the hub pool client's lookback to find events in the l1 to l2 direction, which is hardcoded to 8 hours. But if we wanted to increase this lookback using the MAX_FINALIZER_LOOKBACK environment variable, it would unexpectedly not affect the lookback in the L1 to L2 direction.