fix(relayer): Suppress duplicate event data from external listeners (#2083)
The relayer listener processes are supplying the relayer with duplicate
events. This happens when there are > quorum RPC providers that all 
deliver an identical event to the listener process, but with a block
received by the listener process inbetween. In this case the event is
treated as two separate events because the listener does not record that
it has already submitted the event previously.

This commit introduces basic state tracking to ensure that an event 
already submitted to the relayer is not re-submitted thereafter.