fix(InventoryClient): Handle when ProposedRootBundle event preceding RootBundleExecuted event cannot be found
Its possible that the RootBundleExecued event for a chain and token is old enough that the ProposedRootBundle event is older than the earliest block searched. In this case,  our best way to approximate the running balance is to use the running balance from the RootBundleExecuted event plus any deposits and minus any refunds seen in the entire spoke pool client's lookback. This lookback window is correct to use because the bundle end block associated with the RootBundleExecued by definition must be older than the spoke pool client's lookback. If we can't even see the ProposedRootBundle event then the bundle end block must be even older