improve(InventoryClient): Align allocation percentage computations between Rebalancer and Relayer (#1523)
* WIP: improve(InventoryClient): Align allocation percentage computations between Rebalancer and Relayer

As part of a larger inventory client cleanup/refactor I noticed that there is possibly a bug in how we compute "allocation" percentages when (1) determining repayment chain for fills and (2) computing possible rebalances.

For (2), we subtract the "token shortfall" from the numerator but in (2) we subtract it from both the numerator and the denominator. This is inconsistent and I don't think it makes sense.

A shortfall represents a lack of tokens on a chain suggesting that we need that amount of tokens to fill a deposit. The rebalancer (i.e. function (2) above) it designed to rebalance inventory from L1 to L2 to fulfill these shortfalls. Therefore, it makes sense to me that in (2) the shortfall is subtracted from the numerator to compute the allocation percentage on the L2.

The cumulative balance technically is the total inventory across all chains. This should intuitively include all outstanding transfers between L1 and L2 because this is still part of the inventory, its just in-flight between chains.

A shortfall will lead to an outstanding transfer but it will also lead to a subtraction in the total inventory once the transfer finalizes and part of that new L2 balance is used to make a fill.

However, if the shortfall is subtracted from the denominator in the rebalancer/(2) function, then the amount of tokens bridged over to the L2 will be insufficient to fulfill the shortfall.

This PR makes the allocation % computation consistent between (1) and (2) by removing the shortfall subtraction from the denominator in (2).

* Update InventoryClient.ts

* Update InventoryClient.ts

* Update InventoryClient.ts

* Update InventoryClient.RefundChain.ts

* Update InventoryClient.ts