improve: Reduce transaction submission retries (#2128)
Because there are multiple providers for each chain, reduce the number
of retries on transaction submission to speed-up the case where a
transaction is actually unable to be submitted.

There are two layers of transaction submission at play - one at the 
quorum provider level, and one above. The quorum provider ensures that
transactions are submitted successively to each provider until success
or failure. For most chains, this means that transactions are submitted
to 2 - 3 providers per retry. The upper layer then implements its own 
retry mechanism, currently at 2 retries. In cases where the transaction
is ultimately unable to be submitted (i.e. underpriced replacement),
this results in multiple futile submission attempts.