improve(relayer): Deposit rate-limit follow-up
Limiting to n deposits per loop is only useful when the relayer operates
in looping mode. In single-shot mode it can unfortunately cause the
relayer to ignore deposits that it otherwise might have filled.

Don't make this configurable for now in order to keep things simple. It
might however be a candidate for configuration in future.