improve(finalizer): Don't rely on TokensBridged events to finalize L2 to L1 transfers
This is a step towards generalizing the finalizer so that it can also help finalize withdrawals from L2 to L1 sent outside of the normal Spoke --> Hub process.

This PR builds on top of existing code that handles a special set of EOA addresses that we want to finalize for, so it actually deletes as much logic as it adds.