improve(Dataworker): Only set blockRangeEndBlockBuffer in BundleDataClient
We set `blockRangeEndBlockBuffer` on construction in both the Dataworker and the BundleDataClient and this is redundant as the former always has one of the latter and can infer this value from the latter.

This PR will help us improve setting the `blockRangeEndBlockBuffer` but first we should consolidate where we actually set it in the clients. I envision us dynamically setting `blockRangeEndBlockBuffer` before we construct the `BundleDataClient`.