improve(BaseBridgeAdapter): Remove unused event parsing logic (#2174)
* improve(BaseBridgeAdapter): Remove unused event parsing logic

The motivation behind this PR is to reduce the amount of work it costs to support a new canonical bridge in the BaseBridgeAdapter. Currently this involves adding event querying logic for any L1-->L2 transfers through a bridge contract which involves adding logic to handle two events:

1. Initiating the deposit on L1
2. Receiving the deposit on L2

Despite each bridge event looking different, we attempt to unify the logic (so that we can easily track the status of these deposits across different bridges) by implementing an interface for each bridge:
- `queryL1BridgeInitiationEvents()`
- `queryL2BridgeFinalizationEvents()`

These functions ideally should return the same object type, `BridgeEvent` that we can use to then implement `getOutstandingCrossChainTransfers`, which is used by the rebalancer and monitor. To make this work, we modify all `BridgeEvents` to have the same properties: `from`, `to` and `amount`. The latter property, `amount` is super important to make `getOutstandingCrossChainTransfers` work but the former two, `from` and `to` are not actually used anywhere in the code.

This PR removes `from` and `to` from `BridgeEvent` which marginally reduces the work to add a new event.

* fix

* Fix tests

* Update src/adapter/bridges/ZKStackWethBridge.ts

Co-authored-by: bmzig <57361391+bmzig@users.noreply.github.com>

* Update src/adapter/bridges/ZKStackBridge.ts

Co-authored-by: bmzig <57361391+bmzig@users.noreply.github.com>

* Update src/adapter/bridges/ZKStackBridge.ts

Co-authored-by: bmzig <57361391+bmzig@users.noreply.github.com>

* Update src/adapter/bridges/ZKStackBridge.ts

Co-authored-by: bmzig <57361391+bmzig@users.noreply.github.com>

---------

Co-authored-by: bmzig <57361391+bmzig@users.noreply.github.com>