improve(finalizer): Don't rely on TokensBridged events to finalize L2 to L1 transfers (#2166)
* feat: Support CCTP V2 finalizer, rebalancer and monitor

Removed Linea USDC bridge, changed constants

* finis up draft

* Update ContractAddresses.ts

* Update yarn.lock

* fix package json

* Finish tests

* Get finalizations to work without deriving nonceHash

* Pass in more addresses to l2 to l1 finalizers

* lint

* improve(finalizer): Don't rely on TokensBridged events to finalize L2 to L1 transfers

This is a step towards generalizing the finalizer so that it can also help finalize withdrawals from L2 to L1 sent outside of the normal Spoke --> Hub process.

This PR builds on top of existing code that handles a special set of EOA addresses that we want to finalize for, so it actually deletes as much logic as it adds.

* Update arbStack.ts

* Rename variables, remove some extra code

* Refactor a bit

* Update package

* Update UsdcCCTPBridge.ts