feat: Treat lite chain deposits and forced origin chain repayments similarly
## Relayer and InventoryClient changes

If a deposit's inputToken is not mapped to a PoolRebalanceRoute mapping then the fill must take repayment on origin chain, which is similar to a lite chain deposit.

There is a lot of logic in the Relayer and InventoryClient that deals with lite chain deposits that should therefore also apply to these "forced origin chain repayment" deposits.

## Replacing token equivalency mapping lookups

Additionally, this PR replaces more token matching lookups via PoolRebalanceRoute events with looking up in the TOKEN_SYMBOLS_MAP constants file. This is required to make all bots be able to support deposited tokens that do not map to any L1 token via PoolRebalanceRoutes