fix(ZkStackBridge): Fix getOutstandingTransfers
We are double counting Hub->Spoke transfers because of the way the monitor works.

The `BaseChainAdapter` queries L1 and L2 events for each address passed into it by setting the `fromAddress` and `toAddress` [equal to each other](https://github.com/across-protocol/relayer/blob/0e8b9c143a477d409191c72ccfb8cf9314d03a65/src/adapter/BaseChainAdapter.ts#L356). This works great when tracking EOA transfers but not when sending transfers from Hub to Spoke.

Therefore, each adapter must handle this case and not double count such transfers. The ZkStackBridge is correctly tracking when the `from` and `to` addresses are set to the SpokePool, but not when they are set to the HubPool. When the `from` and `to` address is set to the HubPool, then `queryL1BridgeInitiationEvents` is returning some events but `queryL2BridgeFinalizationEvents` is returning nothing, leading the monitor to think there are outstanding transfers.