improve(BlastFinalizer): Filter out Blast withdrawals
We can attempt to finalize blast withdrawals that are not ready, so we should catch them rather than log a `warn` message.

We currently filter out USDB withdrawals correctly but USDB is special in that USDB is auto withdrawn for us and then we submit a second transaction to finalize them. We filter on this second step correctly.