improve(Relayer): Make inventoryClient.canTakeDestinationChainRepayment more robust (#2221)
* improve(Relayer): Make inventoryClient.canTakeDestinationChainRepayment more robust

Currently our check for whether a destination chain has a valid pool rebalance route for a deposit is flawed because it doesn't check if the output token and input token are mapped to the same L1 token.

This isn't a production in production because the relayer filters out any deposits that involve cross chain swaps before it gets to choosing a repayment chain, but this PR will add more future protection in case this filter gets removed.

* Update InventoryClient.ts

* Update InventoryClient.ts

* Fix check