Revert "Update index.ts (#2229)" (#2257)
This reverts commit 54ff57e.

The key revert here is to avoid the bundling of transactions before
they're being supplied to the MultiCallerClient, because this prevents
the MultiCallerClient from filtering out failing transactions and allows
a single bad transaction to spoil the bundle.

The rationale for reverting this change is that there's currently an
impossible-to-execute message awaiting finalization on Linea. Since the 
message can't be executed on Linea (due to SpokePool v4 upgrade) it can 
never be finalized. The only (apparent) way to be rid of it is to wait
for it to drop out of the finalizer lookback. Without this, the sweeper
finalizer would continue to error for the next ~15 days (due to its long
lookback configuration).