improve: Relocate invalid fill logging to monitor
This logging is currently emitted by SDK implementation, irrespective of
which bot personality is running. The relayer is focused on speed and
the large number of perceived invalid fills is causing slowdowns and
rate-limiting on messages to Slack. Relocate this to the monitor to
ensure that the log messages are emitted at a more sensible rate.