chore: Temporarily permit suppressing elastic chain withdrawals
The finalizer is choking on a Lens USDC withdrawal because the L2 USDC
bridge contract does not confirm to the L2SharedBridge interface
expected by the zksync-ethers package. Matter Labs have been asked for
clarification on how to handle these. The withdrawal has been finalized
manually in the mean time.