refactor: combine the proposer and l1 executor (#2358)
executePoolRebalanceLeaves is called only by the l1 executor to execute a root bundle on mainnet after finalization. Since we need to execute the root bundle out of the hub pool before proposing another root bundle, it makes sense to swap the order and instead run the l1 executor before the proposer.

For expediency, the two processes have been combined here. This means we are loading data from the BundleDataClient in parallel. I believe this should be fine. In practice, one load is being done via arweave data while the other is being loaded via loadDataFromScratch, but even if both pieces are being loaded from scratch, it doesn't seem like they write to any piece of memory that the other could possibly read from.

Signed-off-by: Bennett <bennett@umaproject.org>
Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>