improve(proposer): Guard early exit in proposer/executor
tldr this restores the existing proposer & executor behaviour from
before the recent change to bail early.

RUN_IDENTIFIER is typically only defined in production and it has
previously been used in the relayer to infer whether the bot is running
in a production environment.

Also, irrespective of environment, the proposer should run if both the
propser is enabled and FORCE_PROPOSE is set.