feat(TransactionUtils): Do not send `error` level log in `runTransaction()` (#2416)
* feat(TransactionUtils): Allow caller to suppress failure in `runTransaction()`

Adds a flag `errorOnFailure` to `runTransaction()` with a default value of `false`, allowing caller to suppress any failures if a transaction fails to submit on-chain successfully and log at the `debug` level instead of `error`.

This is designed to be used by the `MulticallerClient` which simulates the transaction  calldata before calling `runTransaction` on-chain. This flag is therefore safe to set to `true` when called by the multicaller.

The motivation for this PR is seeing `error` level logs emitted by `runTransaction` caused by seemingly colliding transactions due to interleaving bot runs. For example, if the executor attempts to execute a leaf at the same time across two different runs, they will both pass simulation in the `MulticallerClient` but due to the latency between simulation transactions and ultimately submitting them, only one of the transactions will succeed while the other one will revert.

* Update TransactionUtils.ts

* fix

* revert

* Add warning about simulating before submitting

* Update TransactionClient.ts

* Update TransactionClient.ts

* Update index.ts