feat(Dataworker#index): Add buffer getChallengeRemaining calls
The current bot seems to attempt to submit a transaction before the challenge period has actually passed, therefore I suspect that there is some mysterious error happening in the `getChallengeRemaining()`'s parallel calls to `getCurrentTime` and `pendingRootBundleProposal`. The challenge remaining returns as 0 but when the multicaller client goes to execute the transactions, empirically we're seeing about once a day a failure related to the challenge period liveness not being over on-chain.

This PR adds more robust logs as well as a buffer of 2 blocks.